Move IsDirectComparable() into validators package
